move code to src dir
