+++ x-pack/packages/ml/form_utils/src/components/form_section_toggle.tsx
