+++ x-pack/packages/ml/form_utils/src/utils/capitalize_first_letter.ts
